requirements.txt
Django==1.11.10
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
pyjade==4.0.0
PyJWT==1.5.3
wheel==0.30.0